add custom hook to separate business logic from the component, delete deprecated funcionalities from useCharacters
